+++ app/components/WebsiteCard.vue
@@ -18,7 +18,7 @@ export default {
   <li>
     <a :href="websiteItem.href" :title="websiteItem.desc" target="_blank" rel="noopener noreferrer" class="card-link" @click="trackClick">
       <div class="card-icon">
-        <span v-if="websiteItem.logo === ''" class="card-avatar">{{ websiteItem.name.charAt(0) }}</span>
+        <span v-if="!websiteItem.logo" class="card-avatar">{{ websiteItem.name.charAt(0) }}</span>
         <img v-else :src="websiteItem.logo" :alt="websiteItem.name" loading="lazy" class="card-logo">
       </div>
       <div class="card-body">